refactor crypt content and lots of clean up
